assesment/code/requirements.txt
pygame-ce
numpy
sys
os